requirements.txt
dropbox==11.36.2
pillow==9.0.1
requests~=2.31.0
